Fix a bug in transit snapping introduced by 163244ace982946f2bddd2fa50745c82e3a3e9d4. Regenerating all maps, also bringing in a Tempe OSM fix...
The deadend trimming is too enthusiastic, getting rid of some unsnapped
cycleways and things connecting to the map border. Will iterate on it
this week; net benefit for now.

+++ importer/actdev_site.sh
@@ -12,28 +12,32 @@ if [ "$SITE" == "" ]; then
 fi
 CITY=`echo $SITE | sed -r 's/-/_/g'`
 
-# Follow https://a-b-street.github.io/docs/user/new_city.html and import as a new city.
-cp -Rv importer/config/gb/leeds importer/config/gb/$CITY
-rm -fv importer/config/gb/$CITY/*.poly
-wget https://raw.githubusercontent.com/cyipt/actdev/main/data-small/$SITE/small-study-area.geojson
-cargo run --bin geojson_to_osmosis < small-study-area.geojson
-rm -fv small-study-area.geojson
-mv boundary0.poly importer/config/gb/$CITY/center.poly
-GEOFABRIK=`cargo run --bin pick_geofabrik importer/config/gb/$CITY/center.poly`
-echo "Geofabrik URL is $GEOFABRIK"
-perl -pi -e "s#\"osm_url\": \".*\"#\"osm_url\": \"$GEOFABRIK\"#" importer/config/gb/$CITY/cfg.json
+# Disable to just regenerate scenarios
+# (TODO But still have to switch the above to handle city as input)
+if false; then
+	# Follow https://a-b-street.github.io/docs/user/new_city.html and import as a new city.
+	cp -Rv importer/config/gb/leeds importer/config/gb/$CITY
+	rm -fv importer/config/gb/$CITY/*.poly
+	wget https://raw.githubusercontent.com/cyipt/actdev/main/data-small/$SITE/small-study-area.geojson
+	cargo run --bin geojson_to_osmosis < small-study-area.geojson
+	rm -fv small-study-area.geojson
+	mv boundary0.poly importer/config/gb/$CITY/center.poly
+	GEOFABRIK=`cargo run --bin pick_geofabrik importer/config/gb/$CITY/center.poly`
+	echo "Geofabrik URL is $GEOFABRIK"
+	perl -pi -e "s#\"osm_url\": \".*\"#\"osm_url\": \"$GEOFABRIK\"#" importer/config/gb/$CITY/cfg.json
 
-wget https://raw.githubusercontent.com/cyipt/actdev/main/data-small/$SITE/site.geojson -O data/system/study_areas/$SITE.geojson
+	wget https://raw.githubusercontent.com/cyipt/actdev/main/data-small/$SITE/site.geojson -O data/system/study_areas/$SITE.geojson
 
-./import.sh --raw --map --city=gb/$CITY
-
-# Procedurally generate houses, if needed
-if cargo run --release --bin generate_houses -- --map=data/system/gb/$CITY/maps/center.bin --num_required=1000 --rng_seed=42 --out=data/input/gb/$CITY/procgen_houses.json; then
-	# Update the importer config, and import again
-	perl -pi -e "s#\"extra_buildings\": null#\"extra_buildings\": \"data/input/gb/$CITY/procgen_houses.json\"#" importer/config/gb/$CITY/cfg.json
 	./import.sh --raw --map --city=gb/$CITY
-else
-	echo "$CITY already had enough houses"
+
+	# Procedurally generate houses, if needed
+	if cargo run --release --bin generate_houses -- --map=data/system/gb/$CITY/maps/center.bin --num_required=1000 --rng_seed=42 --out=data/input/gb/$CITY/procgen_houses.json; then
+		# Update the importer config, and import again
+		perl -pi -e "s#\"extra_buildings\": null#\"extra_buildings\": \"data/input/gb/$CITY/procgen_houses.json\"#" importer/config/gb/$CITY/cfg.json
+		./import.sh --raw --map --city=gb/$CITY
+	else
+		echo "$CITY already had enough houses"
+	fi
 fi
 
 # Import the scenarios
